move useTrackToggle hook
