+++ packages/react/src/components/controls/TrackToggle.tsx
@@ -1,10 +1,7 @@
 import type { CaptureOptionsBySource, ToggleSource } from '@livekit/components-core';
-import { log, setupManualToggle, setupMediaToggle } from '@livekit/components-core';
 import * as React from 'react';
-import { mergeProps } from '../../mergeProps';
-import { useMaybeRoomContext } from '../../context';
 import { getSourceIcon } from '../../assets/icons/util';
-import { useObservableState } from '../../hooks/internal/useObservableState';
+import { useTrackToggle } from '../../hooks';
 
 /** @public */
 export interface TrackToggleProps<T extends ToggleSource>
@@ -14,68 +11,6 @@ export interface TrackToggleProps<T extends ToggleSource>
   initialState?: boolean;
   onChange?: (enabled: boolean) => void;
   captureOptions?: CaptureOptionsBySource<T>;
-}
-
-/** @public */
-export interface UseTrackToggleProps<T extends ToggleSource>
-  extends Omit<TrackToggleProps<T>, 'showIcon'> {}
-
-/** @public */
-export function useTrackToggle<T extends ToggleSource>({
-  source,
-  onChange,
-  initialState,
-  captureOptions,
-  ...rest
-}: UseTrackToggleProps<T>) {
-  const room = useMaybeRoomContext();
-  const track = room?.localParticipant?.getTrack(source);
-
-  const { toggle, className, pendingObserver, enabledObserver } = React.useMemo(
-    () => (room ? setupMediaToggle<T>(source, room, captureOptions) : setupManualToggle()),
-    [room, source, JSON.stringify(captureOptions)],
-  );
-
-  const pending = useObservableState(pendingObserver, false);
-  const enabled = useObservableState(enabledObserver, initialState ?? !!track?.isEnabled);
-
-  React.useEffect(() => {
-    onChange?.(enabled);
-  }, [enabled, onChange]);
-
-  React.useEffect(() => {
-    if (initialState !== undefined) {
-      log.debug('forcing initial toggle state', source, initialState);
-      toggle(initialState);
-    }
-    // only execute once at the beginning
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const newProps = React.useMemo(() => mergeProps(rest, { className }), [rest, className]);
-
-  const clickHandler: React.MouseEventHandler<HTMLButtonElement> = React.useCallback(
-    (evt) => {
-      toggle();
-      rest.onClick?.(evt);
-    },
-    [rest, toggle],
-  );
-
-  return {
-    toggle,
-    enabled,
-    pending,
-    track,
-    buttonProps: {
-      ...newProps,
-      'aria-pressed': enabled,
-      'data-lk-source': source,
-      'data-lk-enabled': enabled,
-      disabled: pending,
-      onClick: clickHandler,
-    } as React.ButtonHTMLAttributes<HTMLButtonElement>,
-  };
 }
 
 /**
