+++ packages/react/src/hooks/index.ts
@@ -1,3 +1,4 @@
+export { useTrackToggle, UseTrackToggleProps } from './useTrackToggle';
 export { useStartAudio, UseStartAudioProps } from './useStartAudio';
 export { useMediaDevices } from './useMediaDevices';
 export { useToggleChat } from './useToggleChat';
