+++ packages/react/src/hooks/useTrackToggle.ts
@@ -1,0 +1,69 @@
+import type { ToggleSource } from '@livekit/components-core';
+import { setupMediaToggle, setupManualToggle, log } from '@livekit/components-core';
+import React from 'react';
+import type { TrackToggleProps } from '../components';
+import { useMaybeRoomContext } from '../context';
+import { mergeProps } from '../mergeProps';
+import { useObservableState } from './internal';
+
+/** @public */
+export interface UseTrackToggleProps<T extends ToggleSource>
+  extends Omit<TrackToggleProps<T>, 'showIcon'> {}
+
+/** @public */
+export function useTrackToggle<T extends ToggleSource>({
+  source,
+  onChange,
+  initialState,
+  captureOptions,
+  ...rest
+}: UseTrackToggleProps<T>) {
+  const room = useMaybeRoomContext();
+  const track = room?.localParticipant?.getTrack(source);
+
+  const { toggle, className, pendingObserver, enabledObserver } = React.useMemo(
+    () => (room ? setupMediaToggle<T>(source, room, captureOptions) : setupManualToggle()),
+    [room, source, JSON.stringify(captureOptions)],
+  );
+
+  const pending = useObservableState(pendingObserver, false);
+  const enabled = useObservableState(enabledObserver, initialState ?? !!track?.isEnabled);
+
+  React.useEffect(() => {
+    onChange?.(enabled);
+  }, [enabled, onChange]);
+
+  React.useEffect(() => {
+    if (initialState !== undefined) {
+      log.debug('forcing initial toggle state', source, initialState);
+      toggle(initialState);
+    }
+    // only execute once at the beginning
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  const newProps = React.useMemo(() => mergeProps(rest, { className }), [rest, className]);
+
+  const clickHandler: React.MouseEventHandler<HTMLButtonElement> = React.useCallback(
+    (evt) => {
+      toggle();
+      rest.onClick?.(evt);
+    },
+    [rest, toggle],
+  );
+
+  return {
+    toggle,
+    enabled,
+    pending,
+    track,
+    buttonProps: {
+      ...newProps,
+      'aria-pressed': enabled,
+      'data-lk-source': source,
+      'data-lk-enabled': enabled,
+      disabled: pending,
+      onClick: clickHandler,
+    } as React.ButtonHTMLAttributes<HTMLButtonElement>,
+  };
+}
